requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
auditlog3==1.0.1
autobahn==23.6.2
Automat==22.10.0
blinker==1.6.2
boto3==1.26.54
botocore==1.29.54
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
coverage==6.5.0
cryptography==41.0.1
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.2
django-advanced-filters==2.0.0
django-allauth==0.51.0
django-cors-headers==4.0.0
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-filter==22.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-test==0.4030
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Faker==12.0.1
faker-file==0.11.1
Flask==2.3.2
Flask-Cors==4.0.0
hyperlink==21.0.0
idna==3.4
image==1.5.33
include==0.2.2
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Markdown==3.4.1
MarkupSafe==2.1.3
marshmallow==3.19.0
mercantile==1.2.1
mixer==7.2.2
mock==5.0.1
msgpack==1.0.5
numpy==1.24.1
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PyMuPDF==1.22.3
pyOpenSSL==23.2.0
PyPDF2==3.0.1
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
service-identity==23.1.0
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.4
twilio==7.15.4
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.4
zope.interface==6.0
